# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Food Palace Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /zones:
    get:
      operationId: listZones
      tags: [zones]
      summary: List delivery zones
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeliveryZone"
    post:
      operationId: createZone
      tags: [zones]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeliveryZoneInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliveryZone"
  /zones/{id}:
    patch:
      operationId: updateZone
      tags: [zones]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeliveryZoneUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliveryZone"
    delete:
      operationId: deleteZone
      tags: [zones]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  # ---- ORDERS ----
...
components:
  schemas:
...
    DeliveryLocation:
      type: object
      required: [id, zoneId, name]
      properties:
        id:
          type: integer
        zoneId:
          type: integer
        name:
          type: string
    DeliveryZone:
      type: object
      required: [id, name, fee, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        fee:
          type: number
        isActive:
          type: boolean
        locations:
          type: array
          items:
            $ref: "#/components/schemas/DeliveryLocation"
        createdAt:
          type: string
          format: date-time
    DeliveryZoneInput:
      type: object
      required: [name, fee]
      properties:
        name:
          type: string
        fee:
          type: number
        isActive:
          type: boolean
        locations:
          type: array
          items:
            type: string
    DeliveryZoneUpdate:
      type: object
      properties:
        name:
          type: string
        fee:
          type: number
        isActive:
          type: boolean

    # ---- ORDER SCHEMAS ----
...